Relocating functions.
